chore:move modeling code location
