+++ optimum/neuron/__init__.py
@@ -76,7 +76,7 @@ _import_structure = {
 }
 
 # Model structures
-_import_structure["models.bert"] = [
+_import_structure["models.inference.bert"] = [
     "NeuronBertModel",
     "NeuronBertForMaskedLM",
     "NeuronBertForQuestionAnswering",
@@ -84,9 +84,9 @@ _import_structure["models.bert"] = [
     "NeuronBertForTokenClassification",
     "NeuronBertForMultipleChoice",
 ]
-_import_structure["models.clip"] = ["NeuronCLIPModel", "NeuronCLIPForImageClassification"]
-_import_structure["models.whisper"] = ["NeuronWhisperForConditionalGeneration"]
-_import_structure["models.yolos"] = ["NeuronYolosForObjectDetection"]
+_import_structure["models.inference.clip"] = ["NeuronCLIPModel", "NeuronCLIPForImageClassification"]
+_import_structure["models.inference.whisper"] = ["NeuronWhisperForConditionalGeneration"]
+_import_structure["models.inference.yolos"] = ["NeuronYolosForObjectDetection"]
 
 if TYPE_CHECKING:
     from .accelerate import NeuronAccelerator, NeuronAcceleratorState, NeuronPartialState
@@ -127,7 +127,7 @@ if TYPE_CHECKING:
     from .modeling_traced import NeuronTracedModel
 
     # Modeling
-    from .models.bert import (
+    from .models.inference.bert import (
         NeuronBertForMaskedLM,
         NeuronBertForMultipleChoice,
         NeuronBertForQuestionAnswering,
@@ -135,9 +135,9 @@ if TYPE_CHECKING:
         NeuronBertForTokenClassification,
         NeuronBertModel,
     )
-    from .models.clip import NeuronCLIPForImageClassification, NeuronCLIPModel
-    from .models.whisper import NeuronWhisperForConditionalGeneration
-    from .models.yolos import NeuronYolosForObjectDetection
+    from .models.inference.clip import NeuronCLIPForImageClassification, NeuronCLIPModel
+    from .models.inference.whisper import NeuronWhisperForConditionalGeneration
+    from .models.inference.yolos import NeuronYolosForObjectDetection
     from .pipelines import pipeline
     from .trainers import NeuronORPOTrainer, NeuronSFTTrainer, NeuronTrainer, Seq2SeqNeuronTrainer
     from .training_args import NeuronTrainingArguments, Seq2SeqNeuronTrainingArguments
