+++ optimum/neuron/models/inference/bert/__init__.py
@@ -13,4 +13,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .model import NeuronYolosForObjectDetection
+from .modeling_bert import (
+    NeuronBertForMaskedLM,
+    NeuronBertForMultipleChoice,
+    NeuronBertForQuestionAnswering,
+    NeuronBertForSequenceClassification,
+    NeuronBertForTokenClassification,
+    NeuronBertModel,
+)
