+++ optimum/neuron/models/inference/clip/modeling_clip.py
