+++ optimum/neuron/models/inference/whisper/__init__.py
@@ -13,4 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .model import NeuronWhisperForConditionalGeneration
+from .modeling_whisper import NeuronWhisperForConditionalGeneration
