+++ optimum/neuron/models/inference/yolos/__init__.py
@@ -13,11 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .model import (
-    NeuronBertForMaskedLM,
-    NeuronBertForMultipleChoice,
-    NeuronBertForQuestionAnswering,
-    NeuronBertForSequenceClassification,
-    NeuronBertForTokenClassification,
-    NeuronBertModel,
-)
+from .modeling_yolos import NeuronYolosForObjectDetection
