+++ optimum/neuron/models/inference/yolos/modeling_yolos.py
